requirements.txt
Flask==3.0.0
Werkzeug==3.1.3
numpy==1.26.4
pandas==2.1.3
Pillow==10.1.0
tensorflow==2.19.0
openpyxl==3.1.2
gunicorn==21.2.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2 